cluster_maxcut/requirements.txt
qoro-divi
networkx
matplotlib
